Switched to a server-first architecture for faster post retrieval
Now, the client has two ways of getting the list of posts
- (new method) calling out to a server which will build the list of posts
  and send it to the client
- (old method) crawling the API manually

The new method is significantly faster because it reduces the number of
requests and network overhead. The old method is more resilient because it
is only dependent on one server. The client will first check the server,
before falling back to manual crawling.